move wallet-connect page to integrations
