+++ src/pages/integrations/wallet-connect/wallet-connect.html
