+++ src/pages/integrations/wallet-connect/wallet-connect.scss
